refactor(plot): move color_selector.py into util directory
