+++ src/caret_analyze/plot/visualize_lib/bokeh/util/__init__.py
@@ -1,7 +1,9 @@
+from .color_selector import ColorSelectorFactory
 from .legend import HoverCreator, HoverKeys, HoverSource, LegendManager
 from .util import RectValues
 
 __all__ = [
+    "ColorSelectorFactory",
     "HoverCreator",
     "HoverKeys",
     "HoverSource",
